# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "影子简历助手本地 API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/session-check": {
      "get": {
        "summary": "Session Check",
        "operationId": "session_check_api_session_check_get",
        "parameters": [
          {
            "name": "x-shadow-session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Shadow-Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Session Check Api Session Check Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/internal/shutdown": {
      "post": {
        "summary": "Shutdown",
        "operationId": "shutdown_internal_shutdown_post",
        "parameters": [
          {
            "name": "x-shadow-session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Shadow-Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Shutdown Internal Shutdown Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/bootstrap": {
      "get": {
        "summary": "Get Bootstrap",
        "operationId": "get_bootstrap_api_bootstrap_get",
        "parameters": [
          {
            "name": "x-shadow-session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Shadow-Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BootstrapState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update Bootstrap",
        "operationId": "update_bootstrap_api_bootstrap_patch",
        "parameters": [
          {
            "name": "x-shadow-session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Shadow-Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OnboardingUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BootstrapState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/settings": {
      "get": {
        "summary": "Get Settings",
        "operationId": "get_settings_api_settings_get",
        "parameters": [
          {
            "name": "x-shadow-session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Shadow-Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Settings Api Settings Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update Settings",
        "operationId": "update_settings_api_settings_put",
        "parameters": [
          {
            "name": "x-shadow-session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Shadow-Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Settings Api Settings Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/credentials/openai": {
      "put": {
        "summary": "Put Openai Key",
        "operationId": "put_openai_key_api_credentials_openai_put",
        "parameters": [
          {
            "name": "x-shadow-session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Shadow-Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiKeyInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Put Openai Key Api Credentials Openai Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Openai Key",
        "operationId": "delete_openai_key_api_credentials_openai_delete",
        "parameters": [
          {
            "name": "x-shadow-session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Shadow-Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Delete Openai Key Api Credentials Openai Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApiKeyInput": {
        "properties": {
          "api_key": {
            "type": "string",
            "maxLength": 512,
            "minLength": 8,
            "title": "Api Key"
          }
        },
        "type": "object",
        "required": [
          "api_key"
        ],
        "title": "ApiKeyInput"
      },
...
      "BootstrapState": {
        "properties": {
          "privacy_accepted": {
            "type": "boolean",
            "title": "Privacy Accepted"
          },
          "initialized": {
            "type": "boolean",
            "title": "Initialized"
          },
          "onboarding_step": {
            "type": "integer",
            "title": "Onboarding Step"
          },
          "api_key_configured": {
            "type": "boolean",
            "title": "Api Key Configured"
          },
          "data_directory": {
            "type": "string",
            "title": "Data Directory"
          }
        },
        "type": "object",
        "required": [
          "privacy_accepted",
          "initialized",
          "onboarding_step",
          "api_key_configured",
          "data_directory"
        ],
        "title": "BootstrapState"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OnboardingUpdate": {
        "properties": {
          "privacy_accepted": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Privacy Accepted"
          },
          "initialized": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Initialized"
          },
          "onboarding_step": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 4.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Onboarding Step"
          }
        },
        "type": "object",
        "title": "OnboardingUpdate"
      },
...
      "SettingsUpdate": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider",
            "default": "openai"
          },
          "model": {
            "type": "string",
            "title": "Model",
            "default": "gpt-5-mini"
          },
          "api_mode": {
            "type": "string",
            "enum": [
              "responses",
              "chat_completions"
            ],
            "title": "Api Mode",
            "default": "responses"
          },
          "base_url": {
            "type": "string",
            "title": "Base Url",
            "default": ""
          },
          "voice_device_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voice Device Id"
          }
        },
        "type": "object",
        "title": "SettingsUpdate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}